+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 | :--------- | :------- | :---- |
 | Easy       | 21       | 492   |
 | Medium     | 56       | 837   |
-| Hard       | 17       | 328   |
+| Hard       | 18       | 328   |
 
 ## 0004 - Median Of Two Sorted Arrays
 
@@ -368,6 +368,11 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - `nums[:a]` are red, `nums[a:b]` are white, `nums[c:]` are blue
 - `nums[b:c]` are unknown
+
+## 0076 - Minimum Window Substring
+
+- Two points
+- Slice window
 
 ## 0077 - Combinations